fix(blink.cmp): fix doc and signature help colours
They were set too dark and were not visible. The blink.cmp documentation
isn't really very clear to be honest, "BlinkCmpDoc" is actually the hl
group for regular text in the doc window, not the background of the
window, and the same is true for signature help, and for
BlinkCmpMenu.